frontend/bookstore/requirements.txt
django==5.0.0
djangorestframework==3.14.0
requests==2.31.0
pillow==10.0.0
